Moved parsing functions from commands.ts to utils.ts
